% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{estimate_porosity}
\alias{estimate_porosity}
\title{Estimate porosity}
\usage{
estimate_porosity(responses, df)
}
\arguments{
\item{responses}{(DESCRIPTION)}

\item{df}{(DESCRIPTION)}
}
\value{

}
\description{
Estimate porosity
}
