% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{find_dimensional_difference}
\alias{find_dimensional_difference}
\title{Difference along dimensions}
\usage{
find_dimensional_difference(responses_one_dim)
}
\arguments{
\item{responses_one_dim}{(DESCRIPTION)}
}
\value{

}
\description{
Difference along dimensions
}
