% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{get_dimension}
\alias{get_dimension}
\title{Calculate dimensions from coordinate data}
\usage{
get_dimension(df, grid)
}
\arguments{
\item{df}{The image dataframe}

\item{grid}{(DESCRIPTION)}
}
\value{

}
\description{
Calculate dimensions from coordinate data
}
\examples{
(EMPTY)
}
