% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereology_beta_02.R
\name{get_metric}
\alias{get_metric}
\title{Recalculate porosity and SE}
\usage{
get_metric(object)
}
\arguments{
\item{dataframe}{object of image}
}
\value{

}
\description{
Recalculate porosity and SE
}
\examples{
get_metric(df)
}
