% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{img_to_table}
\alias{img_to_table}
\title{Convert image to dataframe}
\usage{
img_to_table(img, dimx = 300, dimy = 300)
}
\arguments{
\item{img}{(DESCRIPTION)}

\item{dimx}{(DESCRIPTION)}

\item{dimy}{It reduces the size of the data at the same type to a specified dimension (dimx, dimy)}
}
\value{

}
\description{
Convert image to dataframe
}
