% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereology_beta_02.R
\name{line_estimator_app}
\alias{line_estimator_app}
\title{Porosity estimator}
\usage{
line_estimator_app(
  object = NULL,
  path = system.file("extdata", "sponge3.jpg", package = "Stereology"),
  grid_number = 10,
  seed = 1
)
}
\arguments{
\item{path}{}
}
\value{
Data table with locations of crossings
}
\description{
The function line_estimator_app() takes several arguments: object (default = NULL) is an image data frame in long format with columns x, y, and hexacode for color; path is the path to an image, and can be used instead of the object; grid_number decides the vertical lines (horisontal lines are separated by the same distance as the x-axis lines and might differ from the grid_number); and finally seed makes the random initialisation of the gridlines a bit less random.
Rules of the line estimator:
- One must always assign two points to one air bubble.
- If one misses a point, then assign a new point on top (it is not possible to remove points).
- Always make sure lines are completed before using the estimates.
- Alternate between completing horisontal and vertical lines to reduce potential bias and obtain a better estimate.
}
\examples{
line_estimator_app()
}
