% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{make_grid}
\alias{make_grid}
\title{Initiate grid matching image dimensions}
\usage{
make_grid(img_tabledata, n = 10)
}
\arguments{
\item{img_tabledata}{(DESCRIPTION)}

\item{n}{(DESCRIPTION)}
}
\value{

}
\description{
Random start point of grid.
}
