% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{make_grid_plot}
\alias{make_grid_plot}
\title{Plot data and grid}
\usage{
make_grid_plot(img_tabledata, grid)
}
\arguments{
\item{img_tabledata}{(DESCRIPTION)}

\item{grid}{(DESCRIPTION)}
}
\value{

}
\description{
Plot data and grid
}
