% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretic_part.R
\name{mat.gen}
\alias{mat.gen}
\title{Theoretic estimations.}
\usage{
mat.gen(image_path)
}
\arguments{
\item{image_path}{the path where the image is stored.}

\item{mtr}{a matrix generated using the function 'mat.gen()'.}

\item{x}{the index of the pixel-column to subset the 'mtr'.}

\item{d}{the length used to subset the 'mtr'.}

\item{trial_s}{amount of repetitions for the simulation.}
}
\description{
2 functions ('mat.gen()' and 'th.i.est()') offering exploretion of theoretical estimates of an image. For more details use: browseVignettes(package = "Stereology")
}
\examples{
# Loading the image.

# To use a PNG/JPEG/BMP image stored on your computer try:
# path <- "~/images.png"
path <- system.file('extdata/sponge3.jpg',package='Stereology')

# Applying mat.gen().
mat <- mat.gen(path)

# Applying thr.i.ests()
est.list <- th.i.ests(mtr = mat, x = 10, y = 10, d= 10, trial_s = 1000)
est.list

# For more details use: browseVignettes(package = "Stereology").
}
