% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{plot.stero}
\alias{plot.stero}
\title{Plot stero image}
\usage{
\method{plot}{stero}(x)
}
\arguments{
\item{stero}{object}
}
\value{
A ggplot object visualizing the loaded picture
}
\description{
Plot stero image
}
\examples{
plot(img)
}
