% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_functions.R
\name{se_prop}
\alias{se_prop}
\title{Calculate standard error of proportion}
\usage{
se_prop(proportion, pop_size)
}
\arguments{
\item{proportion}{porosity estimate}

\item{pop_size}{number of pairs (i.e., points / 2)}
}
\value{

}
\description{
Calculate standard error of proportion
}
\examples{
se_prop(proportion, pop_size)

}
