% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereology_beta_02.R
\name{ster}
\alias{ster}
\title{Porosity estimator}
\usage{
ster(
  object = NULL,
  path = system.file("extdata", "sponge3.jpg", package = "Stereology"),
  seed = 1
)
}
\arguments{
\item{path}{}
}
\value{
Data table with locations of crossings
}
\description{
Porosity estimator
}
