% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stereology_beta_02.R
\name{write_data}
\alias{write_data}
\title{Write ster data}
\usage{
write_data(data, ...)
}
\arguments{
\item{data}{from spateology curation}

\item{...}{}
}
\value{
a csv file saved in working folder
}
\description{
Write ster data
}
\examples{
None - to be used in run_spat function
}
