% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretic_part.R
\name{Theoretic}
\alias{Theoretic}
\alias{bw_mat}
\alias{th_i_ests}
\title{Theoretic estimations.}
\usage{
bw_mat(image_path, thr = 0.5)

th_i_ests(mtr, x, y, dx, dy, trial_s)
}
\arguments{
\item{image_path}{The path where the image is stored.}

\item{thr}{The threshold used to turn the image in black and white. The default value is 0.5.}

\item{mtr}{A matrix generated using the function 'bw_mat()'.}

\item{x}{Numeric variable. The index of the pixel-row to start subsetting the
'mtr'.}

\item{y}{Numeric variable. The index of the pixel-column to start subsetting
the 'mtr'.}

\item{dx}{Numeric variable. The length of pixels used to subset the 'mtr' on the x-axis.}

\item{dy}{Numeric variable. The length of pixels used to subset the 'mtr' on the y-axis.}

\item{trial_s}{Numeric variable. The amount of repetitions used for the
simulation.}
}
\value{
The function 'bw_mat()' returns a list containing: \cr
  1) the matrix that represents the black and white version of the image that
  was used as input, and \cr
  2) the plot of the black and white version of the image.

The function 'th_i_est()' returns a list containing: \cr
  1) submatrix: part of the original matrix \cr
  2) p: the probability (p) estimated using this sub-matrix, \cr
  3) mean_p_hat: the average of the p_hats generated through simulations, and \cr
  4) variance_p_hat: the variance of those p_hats.
}
\description{
2 functions: 'bw_mat()' and 'th_i_est()', offering exploration of
  theoretical estimates of an image. For more details use: \cr
  browseVignettes(package = "Stereology")
}
\examples{
# Loading the image.

# To use a PNG/JPEG/BMP image stored on your computer try:
# path <- "~/images.png"
path <- system.file('extdata/sponge3.jpg',package='Stereology')

# Applying bw_mat().
a <- bw_mat(path, .5)

# Applying th_i_ests()
est_l <- th_i_ests(mtr = a$matrix, x = 400, y = 100, dx = 10, dy = 10,  trial_s = 1000)
est_l

# For more details use: browseVignettes(package = "Stereology").
}
