% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretic_part.R
\name{Theoretic}
\alias{Theoretic}
\alias{mat_gen}
\alias{th_i_ests}
\title{Theoretic estimations.}
\usage{
mat_gen(image_path)

th_i_ests(mtr, x, y, d, trial_s)
}
\arguments{
\item{image_path}{The path where the image is stored.}

\item{mtr}{A matrix generated using the function 'mat.gen()'.}

\item{x}{Numeric variable. The index of the pixel-row to start subsetting the
'mtr'.}

\item{y}{Numeric variable. The index of the pixel-column to start subsetting
the 'mtr'.}

\item{d}{Numeric variable. the length of pixels used to subset the 'mtr'.}

\item{trial_s}{Numeric variable. The amount of repetitions used for the
simulation.}
}
\value{
The function 'mat_gen()' outputs the matrix that represents the image
  used as input.
  The function 'th_i_est()' a list containing:
  - part of the original matrix
  - the probability (p) estimated using this sub-matrix,
  - the average of the p_hats generated through simulations, and
  - the variance of those p_hats.
}
\description{
2 functions ('mat_gen()' and 'th_i_est()') offering exploration
  of theoretical estimates of an image. For more details use:
  browseVignettes(package = "Stereology")
}
\examples{
# Loading the image.

# To use a PNG/JPEG/BMP image stored on your computer try:
# path <- "~/images.png"
path <- system.file('extdata/sponge3.jpg',package='Stereology')

# Applying mat_gen().
mat <- mat_gen(path)

# Applying th_i_ests()
est_list <- th_i_ests(mtr = mat, x = 10, y = 10, d= 10, trial_s = 1000)
est_list

# For more details use: browseVignettes(package = "Stereology").
}
