% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoretic_part.R
\name{Theoretic}
\alias{Theoretic}
\alias{bw_mat}
\alias{th_i_ests}
\title{Theoretic estimations.}
\usage{
bw_mat(image_path, thr = 0.5)

th_i_ests(mtr, x, y, lx, ly)
}
\arguments{
\item{image_path}{The path where the image is stored.}

\item{thr}{The threshold used to turn the image in black and white. The default value is 0.5.}

\item{mtr}{A matrix generated using the function 'bw_mat()'.}

\item{x}{Numeric variable. The index of the pixel-row to start the grid.}

\item{y}{Numeric variable. The index of the pixel-column to start the grid.}

\item{lx}{Numeric variable. The amount of horizontal indices to be used.}

\item{ly}{Numeric variable. The amount of vertical indices to be used.}
}
\value{
The function 'bw_mat()' returns a list containing: \cr
  1) the matrix that represents the black and white version of the image that
  was used as input, and \cr
  2) the plot of the black and white version of the image.

The function 'th_i_est()' returns a list containing: \cr
  1) point_mean: the mean of black pixels for the point grid method, and \cr
  2) line_mean: the mean of black pixels for the line grid method.
  3) point_var: the variance of black pixels for the point grid method.
  4) v_lines_var: the variance of black pixels using only the vertical lines of the grid.
  5) h_lines_var: the variance of black pixels using only the horizontal lines of the grid.
  6) sum_lines_var: the sum of the variance of of both horizontal and vertical lines of the grid. There is dependence between "vertical variance" and "horizontal variance", so this output is only useful for counterexamples.
}
\description{
2 functions: 'bw_mat()' and 'th_i_est()', offering exploration of
  theoretical estimates of an image. For more details use: \cr
  browseVignettes(package = "Stereology")
}
\examples{
# Loading the image.

# To use a PNG/JPEG/BMP image stored on your computer try:
# path <- "~/images.png"
path <- system.file('extdata/sponge3.jpg',package='Stereology')

# Applying bw_mat().
a <- bw_mat(path)

# Applying th_i_ests()
est_a <- th_i_ests(mtr = a$img_mat, x = 700, y = 500, lx= 18, ly = 12)
est_a

# For more details use: browseVignettes(package = "Stereology").
}
